Move rocketchat client to seperate package
