+++ pkg/rocketchat/client.go
@@ -1,4 +1,4 @@
-package main
+package rocketchat
 
 import (
 	"bytes"
